+++ NEWS.md
@@ -1,3 +1,16 @@
+# utsf 1.2.0
+
+* The estimated forecast accuracy per horizon is also computed
+* Now it is possible to use only 1 lag with additive or multiplicative
+  transformation, if the features are not transformed.
+* Now it is possible to transform only the target (and not the features) 
+  with the multiplicative transformation.
+* An error is produced if a too large autorregresive lag is used.
+* An error is produced in method KNN when k is greater than the size of the
+  training set.
+* A warning is produced when the time series is too short to estimate
+  forecast accuracy.
+
 # utsf 1.1.0
 
 * Improvements in estimation of forecast accuracy with rolling origin evaluation.